Remove duplicate map generation code What's missing: test for all this code. It's not new code.
