+++ furusiya/model/maps/area_map.py
@@ -1,15 +1,8 @@
 import random
 
+from model.maps.map_tile import MapTile
 
 class AreaMap:
-
-    # TODO: these should be the renderer's concern, not the map's.
-    DARK_GROUND_COLOUR = (32, 32, 32)
-    DARK_WALL_COLOUR = (48, 48, 48)
-    LIGHT_GROUND_COLOUR = (128, 128, 128)
-    LIGHT_WALL_COLOUR = (192, 192, 192)
-    GROUND_CHARACTER = '.'
-    WALL_CHARACTER = '#'
 
     def __init__(self, width, height, walkable=False):
         self.tiles = []
@@ -56,33 +49,11 @@ class AreaMap:
 
         entity.x = x
         entity.y = y
-        self.entities.append(entity) 
-
+        self.entities.append(entity)
+    
 def filter_tiles(tiles, filter_callback):
     return [
         (x, y)
         for x, y in tiles
         if filter_callback(x, y)
     ]
-
-
-# TODO: extract into its own file
-class MapTile:
-    def __init__(self, walkable=False):
-        self.is_explored = False
-        self.convert_to_ground()
-
-    def convert_to_wall(self, character=AreaMap.WALL_CHARACTER, colour=AreaMap.LIGHT_WALL_COLOUR, dark_colour=AreaMap.DARK_WALL_COLOUR):
-        self.is_walkable = False
-        self.block_sight = True
-        self._set_character_and_colour(character, colour, dark_colour)
-        
-    def convert_to_ground(self, character=AreaMap.GROUND_CHARACTER, colour=AreaMap.LIGHT_GROUND_COLOUR, dark_colour=AreaMap.DARK_GROUND_COLOUR):
-        self.is_walkable = True
-        self.block_sight = False
-        self._set_character_and_colour(character, colour, dark_colour)
-        
-    def _set_character_and_colour(self, character, colour, dark_colour):
-        self.character = character
-        self.colour = colour
-        self.dark_colour = dark_colour
